create str dunder for all objects
