+++ test/test_construct.py
@@ -1,27 +1,6 @@
 
 from openpile import construct
-
-def test_from_list2x_parse_top_bottom():
-    # check if integer is passed
-    t, b = construct.from_list2x_parse_top_bottom(2)
-    assert t == 2
-    assert b == 2
-    # check if integer is passed
-    t, b = construct.from_list2x_parse_top_bottom(5.0)
-    assert t == 5.0
-    assert b == 5.0
-    # check if valid list is passed
-    t, b = construct.from_list2x_parse_top_bottom([0.4, 50.6])
-    assert t == 0.4
-    assert b == 50.6
-    
-def test_var_to_str():
-    print_out = construct.var_to_str(11)
-    assert print_out == 11
-    print_out = construct.var_to_str([11,21.2])
-    assert print_out == '11-21.2'
-    
-
+from openpile.core.soilmodels import API_clay
 class TestPile:
     def test_main_constructor(self):
         # create a steel and circular pile
@@ -45,7 +24,7 @@ class TestLayer:
                         top=0,
                         bottom=-10,
                         weight=9,
-                        pymodel=construct.APIclay(Su=[30,35], eps50=[0.01, 0.02], Neq=100), 
+                        lateral_model=API_clay(Su=[30,35], eps50=[0.01, 0.02], Neq=100), 
                     )
 
 class TestMesh:
